Move models into separate files.
