+++ PortalTeme.Data/Models/AcademicYear.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace PortalTeme.Data.Models {
+    public class AcademicYear {
+
+        [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        public Guid Id { get; set; }
+
+        [Required]
+        public string Name { get; set; }
+
+    }
+}
